Add support for OAuth login flows.
This commit add support for OAuth.

We do this by adding an extra FxALoginState called `OAuthLinkedState`. We are able to update the access token by pushing new ones into the account. This has a small risk of having the access token expire just at the wrong time. This is an acceptable risk, given that this code is due to be replaced shortly.

There is also an extra method in `FxALoginHelper`. This should make it possible to support both Browser Assertion flow and OAuth flow at the same time.